trying to build CameraAutoBalance as its own library
